agents/pyproject.toml
[project]
name = "agent-aa-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.0.0",
    "langchain_community",
    "langchain_google_vertexai",
    "langgraph",
    "sqlalchemy-bigquery",
    "google-cloud-bigquery-storage",
    "ipykernel",
    "uvicorn>=0.34.2",
]
